requirements.txt
# Backend
fastapi>=0.116.1
uvicorn[standard]>=0.33.0
python-multipart>=0.0.20

# YOLO 
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=8.0.0
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
tensorboard>=2.4.1
seaborn>=0.11.0
pandas
thop  

# Opsional
onnxruntime>=1.19.2
tensorflow>=2.4.1


